Split establishing of client Connections part
This makes more Client modules `http2` independent
